requirements.txt
thespian==2.5.9
slackclient==1.0.0
PyYAML==3.11